+++ CHANGELOG.md
@@ -4,3 +4,11 @@
 - Removed **Command** property, and her properties, has moved to mitim
 - Removed **eslint and prettier** package
 - Removed auto commands
+
+# 1.1.1
+
+- Fixed some bugs
+- Added **TextCommands** parameter to **Client**
+- Removed `setName` and `setRun` in the event class
+- Changed execute function name in the text and application command classes to `setRun`
+- The options in the commands classes has been changed to optional